Split the crate into io and timer features
