+++ .github/workflows/ci.yml
@@ -30,6 +30,9 @@ jobs:
         if: startsWith(matrix.rust, 'nightly')
         run: cargo check -Z features=dev_dep
       - run: cargo test
+      - run: cargo test --no-default-features
+      - run: cargo test --no-default-features --features io
+      - run: cargo test --no-default-features --features timer
 
   # Copied from: https://github.com/rust-lang/stacker/pull/19/files
   windows_gnu:
@@ -98,6 +101,9 @@ jobs:
       - name: Install Rust
         run: rustup update ${{ matrix.rust }} && rustup default ${{ matrix.rust }}
       - run: cargo build
+      - run: cargo build --no-default-features
+      - run: cargo build --no-default-features --features io
+      - run: cargo build --no-default-features --features timer
 
   clippy:
     runs-on: ubuntu-latest
