+++ Cargo.toml
@@ -19,15 +19,21 @@ name = "io"
 harness = false
 
 [dependencies]
-concurrent-queue = "1.2.2"
+cfg-if = "1.0.0"
+concurrent-queue = { version = "1.2.2", optional = true }
 futures-lite = "1.11.0"
 log = "0.4.11"
 once_cell = "1.4.1"
 parking = "2.0.0"
-polling = "2.0.0"
-slab = "0.4.2"
-socket2 = { version = "0.4.2", features = ["all"] }
+polling = { version = "2.0.0", optional = true }
+slab = { version = "0.4.2", optional = true }
+socket2 = { version = "0.4.2", features = ["all"], optional = true }
 waker-fn = "1.1.0"
+
+[features]
+default = ["io", "timer"]
+io = ["polling", "slab", "socket2"]
+timer = ["concurrent-queue"]
 
 [build-dependencies]
 autocfg = "1"
