+++ tests/async.rs
@@ -1,3 +1,5 @@
+#![cfg(feature = "io")]
+
 use std::future::Future;
 use std::io;
 use std::net::{Shutdown, TcpListener, TcpStream, UdpSocket};
@@ -7,7 +9,9 @@ use std::sync::Arc;
 use std::thread;
 use std::time::Duration;
 
-use async_io::{Async, Timer};
+use async_io::Async;
+#[cfg(feature = "timer")]
+use async_io::Timer;
 use futures_lite::{future, prelude::*};
 #[cfg(unix)]
 use tempfile::tempdir;
@@ -82,6 +86,7 @@ fn tcp_peek_read() -> io::Result<()> {
     })
 }
 
+#[cfg(feature = "timer")]
 #[test]
 fn tcp_reader_hangup() -> io::Result<()> {
     future::block_on(async {
@@ -104,6 +109,7 @@ fn tcp_reader_hangup() -> io::Result<()> {
     })
 }
 
+#[cfg(feature = "timer")]
 #[test]
 fn tcp_writer_hangup() -> io::Result<()> {
     future::block_on(async {
@@ -244,6 +250,7 @@ fn uds_send_to_recv_from() -> io::Result<()> {
     })
 }
 
+#[cfg(feature = "timer")]
 #[cfg(unix)]
 #[test]
 fn uds_reader_hangup() -> io::Result<()> {
@@ -262,6 +269,7 @@ fn uds_reader_hangup() -> io::Result<()> {
     })
 }
 
+#[cfg(feature = "timer")]
 #[cfg(unix)]
 #[test]
 fn uds_writer_hangup() -> io::Result<()> {
@@ -285,6 +293,7 @@ fn uds_writer_hangup() -> io::Result<()> {
 // Test that we correctly re-register interests after we've previously been
 // interested in both readable and writable events and then we get only one of
 // those (we need to re-register interest on the other).
+#[cfg(feature = "timer")]
 #[test]
 fn tcp_duplex() -> io::Result<()> {
     future::block_on(async {
