+++ tests/timer.rs
@@ -1,3 +1,5 @@
+#![cfg(feature = "timer")]
+
 use std::future::Future;
 use std::pin::Pin;
 use std::sync::{Arc, Mutex};
